Fix CI and removed clojure tests from CI
Multiple fixes to CI to reflect changes, including automated scripts.

Updates to API will involve a fair number of changes to clojure tests.
Since these tests mainly focus on testing serialization of full methods,
and since the current version of Cornelius is focusing on single
expressions, this shouldn't result in a reduction in confidence in
program quality.